more or less finished with refactoring indexing for batched fingerprint encoding
